backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.6
google-auth==2.37.0
python-multipart==0.0.20
requests==2.32.3
email-validator==2.2.0
bcrypt==3.2.2